removed cleanup callback chain
