+++ test/dummy/app/models/callback_modifier.rb
@@ -1,5 +1,5 @@
 class CallbackModifier < ActiveRecord::Base
-  has_paper_trail
+  has_paper_trail :on => []
 
   def test_destroy
     transaction do
